+++ angular_forms/test/model_test.dart
@@ -31,13 +31,31 @@ void main() {
       });
       group('dirty', () {
         test('should be false after creating a control', () {
-          var c = new Control('value');
-          expect(c.dirty, false);
+          var control = new Control('value');
+          expect(control.dirty, false);
         });
+
         test('should be true after changing the value of the control', () {
-          var c = new Control('value');
-          c.markAsDirty();
-          expect(c.dirty, true);
+          var control = new Control('value');
+          control.markAsDirty();
+          expect(control.dirty, true);
+        });
+
+        test('should reset to pristine', () {
+          var control = new Control('value');
+          control.markAsDirty();
+          expect(control.dirty, true);
+          control.markAsPristine();
+          expect(control.dirty, false);
+        });
+
+        test('should update parent', () {
+          var control = new Control('value');
+          var group = new ControlGroup({'control': control});
+          control.markAsDirty(onlySelf: false);
+          expect(group.dirty, true);
+          control.markAsPristine(updateParent: true);
+          expect(group.dirty, false);
         });
       });
       group('touched', () {
@@ -265,21 +283,45 @@ void main() {
       });
 
       group('dirty', () {
-        Control c;
-        ControlGroup g;
+        Control control;
+        ControlGroup group;
 
         setUp(() {
-          c = new Control('value');
-          g = new ControlGroup({'one': c});
+          control = new Control('value');
+          group = new ControlGroup({'one': control});
         });
 
         test('should be false after creating a control', () {
-          expect(g.dirty, false);
+          expect(group.dirty, false);
         });
 
         test('should be false after changing the value of the control', () {
-          c.markAsDirty();
-          expect(g.dirty, true);
+          control.markAsDirty();
+          expect(group.dirty, true);
+        });
+
+        test('setting pristine should update control', () {
+          control.markAsDirty();
+          group.markAsPristine();
+          expect(control.dirty, false);
+        });
+
+        test('should derive value from children', () {
+          var otherControl = new Control('new value');
+          group.addControl('two', otherControl);
+
+          //Make only one control dirty.
+          control.markAsDirty();
+          expect(group.dirty, true);
+
+          // Make *both* controls dirty, then pristine only one.
+          otherControl.markAsDirty();
+          otherControl.markAsPristine();
+          expect(group.dirty, true);
+
+          // Now, make *both* controls pristine.
+          control.markAsPristine();
+          expect(group.dirty, false);
         });
       });
 
@@ -509,18 +551,42 @@ void main() {
         });
       });
       group('dirty', () {
-        Control c;
-        ControlArray a;
+        Control control;
+        ControlArray array;
         setUp(() {
-          c = new Control('value');
-          a = new ControlArray([c]);
+          control = new Control('value');
+          array = new ControlArray([control]);
         });
         test('should be false after creating a control', () {
-          expect(a.dirty, false);
+          expect(array.dirty, false);
         });
         test('should be false after changing the value of the control', () {
-          c.markAsDirty();
-          expect(a.dirty, true);
+          control.markAsDirty();
+          expect(array.dirty, true);
+        });
+
+        test('setting pristine should update control', () {
+          control.markAsDirty();
+          array.markAsPristine();
+          expect(control.dirty, false);
+        });
+
+        test('should derive value from children', () {
+          var otherControl = new Control('new value');
+          array.push(otherControl);
+
+          //Make only one control dirty.
+          control.markAsDirty();
+          expect(array.dirty, true);
+
+          // Make *both* controls dirty, then pristine only one.
+          otherControl.markAsDirty();
+          otherControl.markAsPristine();
+          expect(array.dirty, true);
+
+          // Now, make *both* controls pristine.
+          control.markAsPristine();
+          expect(array.dirty, false);
         });
       });